Move ShortlyRepeatedCommandThrottler to its own package
